legecy: merge arclet.alconna.model pyi
